requirements.txt
Flask==1.1.4
Flask-CORS
gunicorn==20.1.0
requests==2.25.1
Werkzeug==0.16.1
MarkupSafe==2.0.1
python-dateutil==2.8.2